MCP actions server for client side MCP tools (#10)
This adds an MCP server that allows client side MCP tools to be executed by MCP clients that operate server side 🤯

For our application this means that the files management and the runtime environment which live inside the browser can be exposed to any MCP client.

This will enable agent frameworks that understand MCP tools (like mastra.ai) to be employed as orchestrator in the next step.